Move client-metadata route into a REST controller; use Rest namespace
- Extract the OAuth client-metadata endpoint out of the Admin class into
  Atmosphere\Rest\Client_Metadata_Controller (includes/rest/), a proper
  WP_REST_Controller — the same structure as the pre-publish controller.
  Public controllers live in Atmosphere\Rest, admin-only ones in
  Atmosphere\Rest\Admin (Rest, not REST, matching wordpress-activitypub).
- All controllers are now registered together in
  Atmosphere::register_rest_controllers() on rest_api_init.
- The route stays atmosphere/v1/client-metadata (the OAuth client_id URL
  is an external contract).
- Document the REST controller convention + namespace versioning in
  docs/php-class-structure.md.

Refs #50
