+++ tests/phpunit/tests/rest/class-test-client-metadata-filter.php
@@ -9,13 +9,13 @@
  *
  * @package Atmosphere
  * @group atmosphere
- * @group wp-admin
+ * @group rest
  */
 
-namespace Atmosphere\Tests\WP_Admin;
+namespace Atmosphere\Tests\Rest;
 
 use WP_UnitTestCase;
-use Atmosphere\WP_Admin\Admin;
+use Atmosphere\Rest\Client_Metadata_Controller;
 
 /**
  * Client metadata filter validation tests.
@@ -36,7 +36,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * required OAuth fields are present and the right shape.
 	 */
 	public function test_default_metadata_is_well_formed() {
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertIsArray( $data );
@@ -53,7 +53,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * Regression for Copilot inline finding on class-admin.php:836.
 	 */
 	public function test_scalar_redirect_uris_falls_back_to_default() {
-		$this->setExpectedIncorrectUsage( 'Atmosphere\\WP_Admin\\Admin::serve_client_metadata' );
+		$this->setExpectedIncorrectUsage( 'Atmosphere\\Rest\\Client_Metadata_Controller::get_metadata' );
 
 		\add_filter(
 			'atmosphere_client_metadata',
@@ -63,7 +63,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 			}
 		);
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertIsArray( $data['redirect_uris'] );
@@ -74,7 +74,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * A filter that returns a non-string `client_id` is rejected.
 	 */
 	public function test_non_string_client_id_falls_back_to_default() {
-		$this->setExpectedIncorrectUsage( 'Atmosphere\\WP_Admin\\Admin::serve_client_metadata' );
+		$this->setExpectedIncorrectUsage( 'Atmosphere\\Rest\\Client_Metadata_Controller::get_metadata' );
 
 		\add_filter(
 			'atmosphere_client_metadata',
@@ -84,7 +84,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 			}
 		);
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertIsString( $data['client_id'] );
@@ -94,11 +94,11 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * A filter that returns a non-array entirely is rejected.
 	 */
 	public function test_non_array_filter_return_falls_back_to_default() {
-		$this->setExpectedIncorrectUsage( 'Atmosphere\\WP_Admin\\Admin::serve_client_metadata' );
+		$this->setExpectedIncorrectUsage( 'Atmosphere\\Rest\\Client_Metadata_Controller::get_metadata' );
 
 		\add_filter( 'atmosphere_client_metadata', static fn() => 'string-instead-of-array' );
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertIsArray( $data );
@@ -112,7 +112,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * off-site URI would be a token-leak primitive.
 	 */
 	public function test_offsite_redirect_uri_falls_back_to_default() {
-		$this->setExpectedIncorrectUsage( 'Atmosphere\\WP_Admin\\Admin::serve_client_metadata' );
+		$this->setExpectedIncorrectUsage( 'Atmosphere\\Rest\\Client_Metadata_Controller::get_metadata' );
 
 		\add_filter(
 			'atmosphere_client_metadata',
@@ -122,7 +122,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 			}
 		);
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertStringStartsWith( \admin_url( '', 'https' ), $data['redirect_uris'][0] );
@@ -134,7 +134,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * server would otherwise deliver the OAuth code over cleartext.
 	 */
 	public function test_http_scheme_redirect_uri_falls_back_to_default() {
-		$this->setExpectedIncorrectUsage( 'Atmosphere\\WP_Admin\\Admin::serve_client_metadata' );
+		$this->setExpectedIncorrectUsage( 'Atmosphere\\Rest\\Client_Metadata_Controller::get_metadata' );
 
 		\add_filter(
 			'atmosphere_client_metadata',
@@ -146,7 +146,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 			}
 		);
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertCount( 1, $data['redirect_uris'] );
@@ -159,7 +159,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * not "partial use the valid ones."
 	 */
 	public function test_mixed_valid_invalid_redirect_uris_falls_back_to_default() {
-		$this->setExpectedIncorrectUsage( 'Atmosphere\\WP_Admin\\Admin::serve_client_metadata' );
+		$this->setExpectedIncorrectUsage( 'Atmosphere\\Rest\\Client_Metadata_Controller::get_metadata' );
 
 		\add_filter(
 			'atmosphere_client_metadata',
@@ -172,7 +172,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 			}
 		);
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		// Should be the default single-entry list, not a 2-entry list.
@@ -186,7 +186,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * the parent array because the array itself has one element.
 	 */
 	public function test_empty_string_redirect_uri_falls_back_to_default() {
-		$this->setExpectedIncorrectUsage( 'Atmosphere\\WP_Admin\\Admin::serve_client_metadata' );
+		$this->setExpectedIncorrectUsage( 'Atmosphere\\Rest\\Client_Metadata_Controller::get_metadata' );
 
 		\add_filter(
 			'atmosphere_client_metadata',
@@ -196,7 +196,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 			}
 		);
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertNotEmpty( $data['redirect_uris'][0] );
@@ -211,7 +211,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 	 * @param mixed $bad_entry Entry to inject.
 	 */
 	public function test_non_string_redirect_uri_entry_falls_back_to_default( $bad_entry ) {
-		$this->setExpectedIncorrectUsage( 'Atmosphere\\WP_Admin\\Admin::serve_client_metadata' );
+		$this->setExpectedIncorrectUsage( 'Atmosphere\\Rest\\Client_Metadata_Controller::get_metadata' );
 
 		\add_filter(
 			'atmosphere_client_metadata',
@@ -221,7 +221,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 			}
 		);
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertIsString( $data['redirect_uris'][0] );
@@ -250,7 +250,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 		// Covers both a numeric entity (&#039;) and a named entity (&amp;).
 		\add_filter( 'pre_option_blogname', static fn() => 'Tom &amp; Toni&#039;s blog' );
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertSame( "Tom & Toni's blog (ATmosphere)", $data['client_name'] );
@@ -269,7 +269,7 @@ class Test_Client_Metadata_Filter extends WP_UnitTestCase {
 			}
 		);
 
-		$response = Admin::serve_client_metadata();
+		$response = ( new Client_Metadata_Controller() )->get_metadata();
 		$data     = $response->get_data();
 
 		$this->assertSame( 'Custom Name', $data['client_name'] );
